+++ frontend/app/src/shared/components/form/fields/relationships/generic-relationship.field.tsx
@@ -17,7 +17,7 @@ import {
 import { FormField, FormInput, FormMessage } from "@/shared/components/ui/form";
 import { Input } from "@/shared/components/ui/input";
 import { useState } from "react";
-import { useDefaultParent } from "./get-default-parent.query";
+import { useDefaultParent } from "../../../../../entities/nodes/relationships/domain/get-default-parent.query";
 
 interface GenericOption extends Node {
   id: string;
