+++ frontend/app/src/shared/components/form/fields/relationships/regular-relationship.field.tsx
@@ -12,7 +12,7 @@ import { PoolSelect } from "@/shared/components/inputs/pool-select";
 import { RelationshipInput } from "@/shared/components/inputs/relationship-one";
 import { FormField, FormInput, FormMessage } from "@/shared/components/ui/form";
 import { useState } from "react";
-import { useDefaultParent } from "./get-default-parent.query";
+import { useDefaultParent } from "../../../../../entities/nodes/relationships/domain/get-default-parent.query";
 
 export interface RegularRelationshipFieldProps extends DynamicRelationshipFieldProps {
   parentDisabled?: boolean;
